pricing_tool_backend/requirements.txt
Django==4.2.5
djangorestframework==3.15.0
psycopg2-binary==2.9.10
django-cors-headers==4.0.0
python-dotenv==1.0.0
djangorestframework-simplejwt